Move default policies into a module
